Iterations on scoring. Fixed BITE.
