+++ notebooks/RunExperiments/runners/experiment_runner.py
@@ -1,18 +1,13 @@
 import argparse
 import copy
-import glob
 import os
 import pickle
 import sys
 import time
 import warnings
-from typing import List, Union
-import cloudpickle
+from typing import List, Optional
 
-import matplotlib
-import matplotlib.pyplot as plt
 import numpy as np
-import pandas as pd
 import ray
 from sklearn.model_selection import train_test_split
 
@@ -20,6 +15,7 @@ from sklearn.model_selection import train_test_split
 from causaltune import CausalTune
 from causaltune.data_utils import CausalityDataset
 from causaltune.models.passthrough import passthrough_model
+from experiment_plots import make_filename
 
 # Ensure CausalTune is in the Python path
 root_path = os.path.realpath("../../../../..")  # noqa: E402
@@ -30,7 +26,7 @@ from causaltune.datasets import load_dataset  # noqa: E402
 from causaltune.search.params import SimpleParamService  # noqa: E402
 from causaltune.score.scoring import (
     metrics_to_minimize,  # noqa: E402
-    supported_metrics,  # noqa: E402
+    # noqa: E402
 )
 
 # Configure warnings
@@ -80,7 +76,14 @@ def parse_arguments():
     return parser.parse_args()
 
 
-def get_estimator_list(dataset_name):
+def get_estimator_list(
+    dataset_name,
+    include_patterns: Optional[List[str]] = None,
+    exclude_patterns: Optional[List[str]] = None,
+):
+    assert (
+        include_patterns is None or exclude_patterns is None
+    ), "Cannot specify both include and exclude patterns"
     if "IV" in dataset_name:
         problem = "iv"
     else:
@@ -92,10 +95,24 @@ def get_estimator_list(dataset_name):
         multivalue=False,
     )
     estimator_list = cfg.estimator_names_from_patterns(problem, "all", 1001)
-    return [est for est in estimator_list if "Dummy" not in est]
+
+    out = [est for est in estimator_list if "Dummy" not in est]
+
+    if include_patterns is not None:
+        out = [est for est in out if any(pat in est for pat in include_patterns)]
+
+    if exclude_patterns is not None:
+        out = [est for est in out if not any(pat in est for pat in exclude_patterns)]
+
+    return out
 
 
-def run_experiment(args, dataset_path: str, use_ray: bool):
+def run_experiment(
+    args,
+    estimators: List[str],
+    dataset_path: str,
+    use_ray: bool,
+):
     # Process datasets
     data_sets = {}
     for dataset in args.datasets:
@@ -110,7 +127,6 @@ def run_experiment(args, dataset_path: str, use_ray: bool):
         data_sets[f"{size} {name}"] = load_dataset(file_path)
 
     out_dir = f"../EXPERIMENT_RESULTS_{args.identifier}"
-
     os.makedirs(out_dir, exist_ok=True)
     out_dir = os.path.realpath(os.path.join(out_dir, size))
     os.makedirs(out_dir, exist_ok=True)
@@ -145,7 +161,7 @@ def run_experiment(args, dataset_path: str, use_ray: bool):
         i_run = 1
 
         for dataset_name, cd in data_sets.items():
-            estimators = get_estimator_list(dataset_name)
+
             # Extract case while preserving original string checking logic
             if "KCKP" in dataset_name:
                 case = "KCKP"
@@ -172,7 +188,6 @@ def run_experiment(args, dataset_path: str, use_ray: bool):
                             args.test_size,
                             args.num_samples,
                             args.components_time_budget,
-                            out_dir,
                             out_fn,
                             estimators,
                         )
@@ -185,8 +200,8 @@ def run_experiment(args, dataset_path: str, use_ray: bool):
                         args.test_size,
                         args.num_samples,
                         args.components_time_budget,
-                        out_dir,
                         out_fn,
+                        estimators,
                     )
                     out.append(results)
 
@@ -216,339 +231,13 @@ def run_experiment(args, dataset_path: str, use_ray: bool):
     return out_dir
 
 
-def extract_metrics_datasets(out_dir: str):
-    metrics = set()
-    datasets = set()
-
-    for file in glob.glob(f"{out_dir}/*.pkl"):
-        parts = os.path.basename(file).split("-")
-        metrics.add(parts[0])
-        datasets.add(parts[-1].replace(".pkl", "").replace("_", " "))
-
-    return sorted(list(metrics)), sorted(list(datasets))
-
-
-def make_filename(metric, dataset, i_run):
-    return f"{metric}-run-{i_run}-{dataset.replace(' ', '_')}.pkl"
-
-
-def get_all_test_scores(out_dir, dataset_name):
-    size, ds_type, case = dataset_name.split(" ")
-    all_scores = []
-    for file in glob.glob(f"{out_dir}/*_{ds_type}_{case}.pkl"):
-        with open(file, "rb") as f:
-            results = pickle.load(f)
-            for x in results["all_scores"]:
-                all_scores.append(
-                    {k: v for k, v in x["test"]["scores"].items() if k not in ["values"]}
-                )
-    out = pd.DataFrame(all_scores)
-    return out
-
-
-def generate_plots(
-    out_dir: str,
-    log_scale: Union[List[str], None] = None,
-    upper_bounds: Union[dict, None] = None,
-    lower_bounds: Union[dict, None] = None,
-    font_size=0,
-):
-    if log_scale is None:
-        log_scale = ["energy_distance", "psw_energy_distance", "frobenius_norm"]
-    if upper_bounds is None:
-        upper_bounds = {}  # Use an empty dictionary if None
-    if lower_bounds is None:
-        lower_bounds = {}  # Use an empty dictionary if None
-
-    metrics, datasets = extract_metrics_datasets(out_dir)
-    # Remove 'ate' from metrics
-    metrics = [m for m in metrics if m.lower() != "ate"]
-
-    metric_names = {
-        "psw_frobenius_norm": "PSW\nFrobenius\nNorm",
-        "frobenius_norm": "Frobenius\nNorm",
-        "erupt": "ERUPT",
-        "codec": "CODEC",
-        "auc": "AUC",
-        "qini": "Qini",
-        "bite": "BITE",
-        "policy_risk": "Policy\nRisk",
-        "energy_distance": "Energy\nDistance",
-        "psw_energy_distance": "PSW\nEnergy\nDistance",
-        "norm_erupt": "Normalized\nERUPT",
-    }
-
-    colors = (
-        [matplotlib.colors.CSS4_COLORS["black"]]
-        + list(matplotlib.colors.TABLEAU_COLORS)
-        + [
-            matplotlib.colors.CSS4_COLORS["lime"],
-            matplotlib.colors.CSS4_COLORS["yellow"],
-            matplotlib.colors.CSS4_COLORS["pink"],
-        ]
-    )
-    markers = ["o", "s", "D", "^", "v", "<", ">", "P", "*", "h", "X", "|", "_", "8"]
-
-    # Determine the problem type from the dataset name
-    problem = "iv" if any("IV" in dataset for dataset in datasets) else "backdoor"
-
-    def plot_grid(title):
-        # Use determined problem type instead of hardcoding "backdoor"
-        files = os.listdir(out_dir)
-        all_metrics = sorted(list(set([f.split("-")[0] for f in files])))
-
-        fig, axs = plt.subplots(
-            len(all_metrics), len(datasets), figsize=(20, 5 * len(all_metrics)), dpi=300
-        )
-
-        if len(all_metrics) == 1 and len(datasets) == 1:
-            axs = np.array([[axs]])
-        elif len(all_metrics) == 1 or len(datasets) == 1:
-            axs = axs.reshape(-1, 1) if len(datasets) == 1 else axs.reshape(1, -1)
-
-        # For multiple metrics in args.metrics, use the first one that has a results file
-        results_files = {}
-        for dataset in datasets:
-            for metric in all_metrics:
-                filename = make_filename(metric, dataset, 1)
-                filepath = os.path.join(out_dir, filename)
-                if os.path.exists(filepath):
-                    results_files[dataset] = filepath
-                    break
-            if dataset not in results_files:
-                print(f"No results file found for dataset {dataset}")
-
-        for j, dataset in enumerate(datasets):
-            if dataset not in results_files:
-                continue
-
-            with open(results_files[dataset], "rb") as f:
-                results = pickle.load(f)
-
-            print(f"Loading results for Dataset: {dataset}")
-
-            for i, metric in enumerate(all_metrics):
-                ax = axs[i, j]
-
-                try:
-                    # Find best estimator for this metric
-                    best_estimator = None
-                    best_score = float("inf") if metric in metrics_to_minimize() else float("-inf")
-                    estimator_name = None
-
-                    for score in results["all_scores"]:
-                        if "test" in score and metric in score["test"]["scores"]:
-                            current_score = score["test"]["scores"][metric]
-                            if metric in metrics_to_minimize():
-                                if current_score < best_score:
-                                    best_score = current_score
-                                    best_estimator = score
-                                    estimator_name = score["test"]["scores"]["estimator_name"]
-                            else:
-                                if current_score > best_score:
-                                    best_score = current_score
-                                    best_estimator = score
-                                    estimator_name = score["test"]["scores"]["estimator_name"]
-
-                    if best_estimator:
-                        CATE_gt = np.array(best_estimator["test"]["CATE_groundtruth"]).flatten()
-                        CATE_est = np.array(best_estimator["test"]["CATE_estimate"]).flatten()
-
-                        # Plotting
-                        ax.scatter(CATE_gt, CATE_est, s=40, alpha=0.5)
-                        ax.plot(
-                            [min(CATE_gt), max(CATE_gt)],
-                            [min(CATE_gt), max(CATE_gt)],
-                            "k-",
-                            linewidth=1.0,
-                        )
-
-                        # Calculate correlation coefficient
-                        corr = np.corrcoef(CATE_gt, CATE_est)[0, 1]
-
-                        # Add correlation
-                        ax.text(
-                            0.05,
-                            0.95,
-                            f"Corr: {corr:.2f}",
-                            transform=ax.transAxes,
-                            verticalalignment="top",
-                            fontsize=font_size + 12,
-                            fontweight="bold",
-                        )
-
-                        # Add estimator name at bottom center
-                        if estimator_name:
-                            estimator_base = estimator_name.split(".")[-1]
-                            ax.text(
-                                0.5,
-                                0.02,
-                                estimator_base,
-                                transform=ax.transAxes,
-                                horizontalalignment="center",
-                                color="blue",
-                                fontsize=font_size + 10,
-                            )
-
-                except Exception as e:
-                    print(f"Error processing metric {metric} for dataset {dataset}: {e}")
-                    ax.text(
-                        0.5,
-                        0.5,
-                        "Error processing data",
-                        ha="center",
-                        va="center",
-                        fontsize=font_size + 12,
-                    )
-
-                if j == 0:
-                    # Create tight layout for ylabel
-                    ax.set_ylabel(
-                        metric_names.get(metric, metric),
-                        fontsize=font_size + 12,
-                        fontweight="bold",
-                        labelpad=5,  # Reduce padding between label and plot
-                    )
-                if i == 0:
-                    ax.set_title(dataset, fontsize=font_size + 14, fontweight="bold", pad=15)
-                ax.set_xticks([])
-                ax.set_yticks([])
-
-        plt.suptitle(
-            f"Estimated CATEs vs. True CATEs: {title}",
-            fontsize=font_size + 18,
-            fontweight="bold",
-        )
-        # Adjust spacing between subplots
-        plt.tight_layout(rect=[0.1, 0, 1, 0.96], h_pad=1.0, w_pad=0.5)
-        plt.savefig(os.path.join(out_dir, "CATE_grid.pdf"), format="pdf", bbox_inches="tight")
-        plt.savefig(os.path.join(out_dir, "CATE_grid.png"), format="png", bbox_inches="tight")
-        plt.close()
-
-    def plot_mse_grid(title):
-        df = get_all_test_scores(out_dir, datasets[0])
-        est_names = sorted(df["estimator_name"].unique())
-
-        # Problem type already determined at top level
-        all_metrics = [
-            c
-            for c in df.columns
-            if c in supported_metrics(problem, False, False) and c.lower() != "ate"
-        ]
-
-        fig, axs = plt.subplots(
-            len(all_metrics), len(datasets), figsize=(20, 5 * len(all_metrics)), dpi=300
-        )
-
-        # Handle single plot cases
-        if len(all_metrics) == 1 and len(datasets) == 1:
-            axs = np.array([[axs]])
-        elif len(all_metrics) == 1 or len(datasets) == 1:
-            axs = axs.reshape(-1, 1) if len(datasets) == 1 else axs.reshape(1, -1)
-
-        legend_elements = []
-        for j, dataset in enumerate(datasets):
-            df = get_all_test_scores(out_dir, dataset)
-            # Apply bounds filtering
-            for m, value in upper_bounds.items():
-                if m in df.columns:
-                    df = df[df[m] < value].copy()
-            for m, value in lower_bounds.items():
-                if m in df.columns:
-                    df = df[df[m] > value].copy()
-
-            for i, metric in enumerate(all_metrics):
-                ax = axs[i, j]
-                this_df = df[["estimator_name", metric, "MSE"]].dropna()
-                this_df = this_df[~np.isinf(this_df[metric].values)]
-
-                if len(this_df):
-                    for idx, est_name in enumerate(est_names):
-                        df_slice = this_df[this_df["estimator_name"] == est_name]
-                        if "Dummy" not in est_name and len(df_slice):
-                            marker = markers[idx % len(markers)]
-                            ax.scatter(
-                                df_slice["MSE"],
-                                df_slice[metric],
-                                color=colors[idx],
-                                s=50,
-                                marker=marker,
-                                linewidths=0.5,
-                            )
-                            if metric not in metrics_to_minimize():
-                                ax.invert_yaxis()
-
-                            trimmed_est_name = est_name.split(".")[-1]
-                            if i == 0 and j == 0:
-                                legend_elements.append(
-                                    plt.Line2D(
-                                        [0],
-                                        [0],
-                                        color=colors[idx],
-                                        marker=marker,
-                                        label=trimmed_est_name,
-                                        linestyle="None",
-                                        markersize=6,
-                                    )
-                                )
-
-                    ax.set_xscale("log")
-                    if metric in log_scale:
-                        ax.set_yscale("log")
-                    ax.grid(True)
-                else:
-                    ax.text(
-                        0.5,
-                        0.5,
-                        "No data",
-                        ha="center",
-                        va="center",
-                        fontsize=font_size + 12,
-                    )
-
-                if j == 0:
-                    # Match ylabel style with plot_grid
-                    ax.set_ylabel(
-                        metric_names.get(metric, metric),
-                        fontsize=font_size + 12,
-                        fontweight="bold",
-                        labelpad=5,
-                    )
-                if i == 0:
-                    ax.set_title(dataset, fontsize=font_size + 14, fontweight="bold", pad=15)
-
-        plt.suptitle(
-            f"MSE vs. Scores: {title}",
-            fontsize=font_size + 18,
-            fontweight="bold",
-        )
-
-        # Match spacing style with plot_grid
-        plt.tight_layout(rect=[0.1, 0, 1, 0.96], h_pad=1.0, w_pad=0.5)
-
-        fig_legend, ax_legend = plt.subplots(figsize=(6, 6))
-        ax_legend.legend(handles=legend_elements, loc="center", fontsize=10)
-        ax_legend.axis("off")
-
-        plt.savefig(os.path.join(out_dir, "MSE_grid.pdf"), format="pdf", bbox_inches="tight")
-        plt.savefig(os.path.join(out_dir, "MSE_grid.png"), format="png", bbox_inches="tight")
-        plt.close()
-
-        # # Create separate legend
-        # fig_legend, ax_legend = plt.subplots(figsize=(6, 6))
-        # ax_legend.legend(handles=legend_elements, loc="center", fontsize=10)
-        # ax_legend.axis("off")
-        # plt.savefig(os.path.join(out_dir, "MSE_legend.pdf"), format="pdf", bbox_inches="tight")
-        # plt.savefig(os.path.join(out_dir, "MSE_legend.png"), format="png", bbox_inches="tight")
-        # plt.close()
-
-    # Generate plots
-    plot_grid("Experiment Results")
-    plot_mse_grid("Experiment Results")
-
-
 def run_batch(
-    identifier: str, kind: str, metrics: List[str], dataset_path: str, use_ray: bool = False
+    identifier: str,
+    kind: str,
+    metrics: List[str],
+    estimators: List[str],
+    dataset_path: str,
+    use_ray: bool = False,
 ):
     args = parse_arguments()
     args.identifier = identifier
@@ -566,11 +255,13 @@ def run_batch(
         # Assuming we port-mapped already by running ray dashboard
         ray.init(
             "ray://localhost:10001",
-            runtime_env={"working_dir": ".", "pip": ["causaltune", "catboost"]},
+            runtime_env={"working_dir": ".", "pip": ["causaltune", "catboost", "ray[tune]"]},
             namespace=RAY_NAMESPACE,
         )
 
-    out_dir = run_experiment(args, dataset_path=dataset_path, use_ray=use_ray)
+    out_dir = run_experiment(
+        args, estimators=estimators, dataset_path=dataset_path, use_ray=use_ray
+    )
     return out_dir
 
 
@@ -617,7 +308,6 @@ def single_run(
     test_size: float,
     num_samples: int,
     components_time_budget: int,
-    out_dir: str,
     out_fn: str,
     estimators: List[str],
     outcome_model: str = "auto",
