+++ notebooks/RunExperiments/runners/kckp_no_meta.py
@@ -1,10 +1,10 @@
 import os
 
-from experiment_runner import run_batch
+from experiment_runner import run_batch, get_estimator_list
 from experiment_plots import generate_plots
 
-identifier = "Egor_test"
-kind = "KC"
+identifier = "Egor_test_no_meta"
+kind = "KCKP"
 metrics = [
     "erupt",
     # "greedy_erupt",  # regular erupt was made probabilistic,
@@ -16,9 +16,17 @@ metrics = [
     "codec",  # NEW
     "bite",  # NEW
 ]
+
+estimators = get_estimator_list(kind, exclude_patterns=["SLearner", "TLearner", "XLearner"])
+
 use_ray = True
 out_dir = run_batch(
-    identifier, kind, metrics, dataset_path=os.path.realpath("../RunDatasets"), use_ray=use_ray
+    identifier,
+    kind,
+    metrics,
+    estimators=estimators,
+    dataset_path=os.path.realpath("../RunDatasets"),
+    use_ray=use_ray,
 )
 # plot results
 # upper_bounds = {"MSE": 1e2, "policy_risk": 0.2}
