+++ notebooks/RunExperiments/runners/rct.py
@@ -1,9 +1,11 @@
 import os
 
-from experiment_runner import run_batch, generate_plots
+from experiment_runner import run_batch, get_estimator_list
+from experiment_plots import generate_plots
 
 identifier = "Egor_test"
 kind = "RCT"
+
 metrics = [
     "erupt",
     # "greedy_erupt",  # regular erupt was made probabilistic,
@@ -15,9 +17,15 @@ metrics = [
     "codec",  # NEW
     "bite",  # NEW
 ]
-use_ray = True
+estimators = get_estimator_list(kind)
+use_ray = False
 out_dir = run_batch(
-    identifier, kind, metrics, dataset_path=os.path.realpath("../RunDatasets"), use_ray=use_ray
+    identifier,
+    kind,
+    metrics,
+    estimators=estimators,
+    dataset_path=os.path.realpath("../RunDatasets"),
+    use_ray=use_ray,
 )
 # plot results
 # upper_bounds = {"MSE": 1e2, "policy_risk": 0.2}
